+++ embassy-stm32/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: stm32/adc/v3: allow DMA reads to loop through enable channels
 - fix: Fix XSPI not disabling alternate bytes when they were previously enabled
 - feat: stm32/fdcan add fn to enable TX buffer mode
+- fix: Fix stm32/fdcan repeatedly reporting the same error
 
 ## 0.3.0 - 2025-08-12